Day 3 part 2 done
Tried to do this in one long chaining command, but without lookahead and
lookbehind features (not existing in Rust's Regex crate), it was found
to be too challenging. Instead, separate the instructions and the parts,
and map all do() instructions into a str that can then be mul'd as in
part1.

Additionally, moved the closure parse_int in sum_pairs out of the loop,
it made little sense to redefine it on every iteration.