Day 5 part 2 use of smart pointer
Previously each page sequence vector would be copied, and then only the
vectors that had a rule broken and were swapped would be accounted for
in the sum. This was a needless waste to clone a vector just to throw it
away. Found a way to do it via smart pointers cow (clone-on-write).
Changed the method to use it for a more efficient approach.