Day 6 part 2 done with brute force
Due to time constraints day 6 part 2 is now done in an inefficient
implementation. It attempts to add an obstacle to every index in the
matrix and then attempts to move, noticing a repeat by being in the same
position as before with the same direction. This works but it runs for a
while, added a TODO to improve this with a graph algorithm once time
permits.